backend/requirements.txt
Flask
Flask-Mail
Flask-CORS
python-dotenv
authlib
PyJWT
mongoengine
user_agents
requests